fix(image-export): own capture serialization in export service
