+++ src/renderer/components/chat/messages/__tests__/MessageList.test.tsx
@@ -1,10 +1,9 @@
+import { captureScrollableAsDataUrl } from '@renderer/services/ExportService'
 import { act, fireEvent, render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import i18n from 'i18next'
 import type { HTMLAttributes, ReactNode, Ref } from 'react'
 import { afterAll, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest'
-
-import { captureScrollable, captureScrollableAsDataUrl } from '@renderer/utils/image'
 
 import { ChatBottomOverlayInsetProvider } from '../../layout/ChatViewportInsetContext'
 import type { MessageVirtualListHandle } from '../list/MessageVirtualList'
@@ -95,8 +94,7 @@ vi.mock('@renderer/hooks/useTimer', () => ({
   })
 }))
 
-vi.mock('@renderer/utils/image', () => ({
-  captureScrollable: vi.fn(),
+vi.mock('@renderer/services/ExportService', () => ({
   captureScrollableAsDataUrl: vi.fn()
 }))
 
@@ -352,7 +350,6 @@ describe('MessageList', () => {
     scrollToKey.mockReset()
     scrollToElement.mockClear()
     scrollToRange.mockClear()
-    vi.mocked(captureScrollable).mockReset()
     vi.mocked(captureScrollableAsDataUrl).mockReset()
     messageVirtualListMocks.deferScrollContainerReady = false
     messageVirtualListMocks.navigationBaseKey = null
